requirements.txt
asttokens==2.4.0
backcall==0.2.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
flake8==6.1.0
fonttools==4.43.1
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.3.3
traitlets==5.11.2
wcwidth==0.2.8
